Convert transactionish to sequence transactions
